blogify/src/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-autoslug==1.9.9
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
django-rosetta==0.10.0
django-tailwind==3.8.0
idna==3.6
jdatetime==4.1.1
phonenumbers==8.13.32
pillow==10.2.0
polib==1.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
